Attempt to implement BlazorWorker, putting this project on hold for now

+++ WarlockSimulatorTBC/Shared/Classes/Spell.cs
@@ -154,7 +154,7 @@ namespace WarlockSimulatorTBC.Shared.Classes
 
 		public override double CalculateCastTime()
 		{
-			return 3 - TalentTree.CurrentTalents["bane"] / 10.0 + player.spellDelay;
+			return 3 - TalentTree.CurrentTalents["bane"] / 10.0 + (double)player.spellDelay;
 		}
 	}
